pyproject.toml
[project]
name = "corethink-mcp"
version = "0.1.0"
description = "A GSR-powered MCP server for safe, natural language reasoning in code changes"
requires-python = ">=3.11.12,<3.12"
dependencies = [
    "mcp[cli]>=0.2.0",
    "fastmcp>=0.1.3",
    "pyyaml>=6.0",
    "gitpython>=3.1.43",
    "python-dotenv>=1.0.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=7.0",
    "pytest-asyncio",
    "black",
    "ruff"
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"corethink_mcp" = ["constraints.txt", "logs/*.log"]

[project.scripts]
corethink-mcp = "corethink_mcp.server.corethink_server:app"
